refac(circom): move `Constraint` to `constraint.h`
